Don't start OAuth flow when moving a to another team w/o integration
Assume for now in the UI that a user has set up the integration for the
source team and that we can just transfer it over to the target team.
